requirements.txt
google-adk>=0.1.0
elevenlabs>=2.0.0
ffmpeg-python>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
websockets>=11.0.3
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
